backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contester-backend"
version = "0.1.0"
description = "Backend API for the local contest management system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask>=3.0,<4.0",
    "Flask-Login>=0.6,<1.0",
    "Flask-SQLAlchemy>=3.1,<4.0",
    "Flask-Migrate>=4.0,<5.0",
    "psycopg[binary]>=3.2,<4.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.6,<1.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"